+++ rust/operator-binary/src/connect/server.rs
@@ -24,7 +24,7 @@ use stackable_operator::{
             apps::v1::{StatefulSet, StatefulSetSpec},
             core::v1::{
                 ConfigMap, EnvVar, HTTPGetAction, PodSecurityContext, Probe, Service,
-                ServiceAccount, ServicePort, ServiceSpec,
+                ServiceAccount,
             },
         },
         apimachinery::pkg::{apis::meta::v1::LabelSelector, util::intstr::IntOrString},
@@ -38,6 +38,7 @@ use stackable_operator::{
 use super::crd::CONNECT_APP_NAME;
 use crate::{
     connect::{
+        GRPC, HTTP,
         common::{self, SparkConnectRole, object_name},
         crd::{
             CONNECT_GRPC_PORT, CONNECT_UI_PORT, DEFAULT_SPARK_CONNECT_GROUP_NAME,
@@ -56,9 +57,6 @@ use crate::{
     },
     product_logging,
 };
-
-const GRPC: &str = "grpc";
-const HTTP: &str = "http";
 
 #[derive(Snafu, Debug)]
 #[allow(clippy::enum_variant_names)]
@@ -393,63 +391,6 @@ pub(crate) fn build_stateful_set(
             ..StatefulSetSpec::default()
         }),
         ..StatefulSet::default()
-    })
-}
-
-// This is the headless driver service used for the internal
-// communication with the executors as recommended by the Spark docs.
-pub(crate) fn build_internal_service(
-    scs: &v1alpha1::SparkConnectServer,
-    app_version_label: &str,
-) -> Result<Service, Error> {
-    let service_name = format!(
-        "{cluster}-{role}-headless",
-        cluster = scs.name_any(),
-        role = SparkConnectRole::Server
-    );
-
-    let selector =
-        Labels::role_selector(scs, CONNECT_APP_NAME, &SparkConnectRole::Server.to_string())
-            .context(LabelBuildSnafu)?
-            .into();
-
-    Ok(Service {
-        metadata: ObjectMetaBuilder::new()
-            .name_and_namespace(scs)
-            .name(service_name)
-            .ownerreference_from_resource(scs, None, Some(true))
-            .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(common::labels(
-                scs,
-                app_version_label,
-                &SparkConnectRole::Server.to_string(),
-            ))
-            .context(MetadataBuildSnafu)?
-            .with_label(Label::try_from(("prometheus.io/scrape", "true")).context(LabelBuildSnafu)?)
-            .build(),
-        spec: Some(ServiceSpec {
-            type_: Some("ClusterIP".to_owned()),
-            cluster_ip: Some("None".to_owned()),
-            ports: Some(vec![
-                ServicePort {
-                    name: Some(String::from(GRPC)),
-                    port: CONNECT_GRPC_PORT,
-                    ..ServicePort::default()
-                },
-                ServicePort {
-                    name: Some(String::from(HTTP)),
-                    port: CONNECT_UI_PORT,
-                    ..ServicePort::default()
-                },
-            ]),
-            selector: Some(selector),
-            // The flag `publish_not_ready_addresses` *must* be `true` to allow for readiness
-            // probes. Without it, the driver runs into a deadlock beacuse the Pod cannot become
-            // "ready" until the Service is "ready" and vice versa.
-            publish_not_ready_addresses: Some(true),
-            ..ServiceSpec::default()
-        }),
-        status: None,
     })
 }